chore: consume auth header
